+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Breaking changes
 * The traits `RegisterFn` and `RegisterResultFn` are removed.  `Engine::register_fn` and `Engine::register_result_fn` are now implemented directly on `Engine`.
 * `FnPtr::call_dynamic` now takes `&NativeCallContext` instead of consuming it.
 * All `Module::set_fn_XXX` methods are removed, in favor of `Module::set_native_fn`.
+* `protected`, `super` are now reserved keywords.
 
 Enhancements
 ------------